Revert to define_method over eval'ing strings
This is much cleaner and requires less code.

Eval was a last resort to try and enable ractor support. It appeared to work with tests passing, but the use of overwritten methods, constants, and the general experimental nature of ractors meant that this library never worked well within ractors in practice.

Keeps and migrates additions for making methods private (#5).

Ractor support will have to wait for ractors to be more mature and may never come if DSL-heavy tools can't be supported by ractors.

Bump version to 1.3.0