Better handling of default values
Previously not specifying a default value meant the option was initialized in the options hash with a nil value.

With this change, the option will no longer be initialized in the options hash if a default was not specified. Checks like `options.key?(:option_name)` will return false for options configured with no default (ie: `option :option_name`).

Option readers will still return `nil` for options with no default specified that haven't been set.

This change is enabled by introducing a default object (`OptStruct::DEFAULT`) as the "default default". If this object is detected as the default for an option, then initialization of the option is skipped. Since the options hash will still return `nil` for keys that haven't been initialized, option readers still return nil.