Fix callback inheritance
During inheritance, struct details are copied to the child by calling `dup` on each instance variable. This works for all of the details, except callbacks.
Those are contained in a `Hash`, grouped by the callback type. Calling `dup` on the `@_callbacks` hash results in a new hash,
but each value in the hash points to the same array as it's parent since the array values within the hash are not duplicated.
This causes new callbacks to be added to both the parent class and the subclass, as well as be shared by all sub-classes of a single parent.

This commit changes the inherited behavior to handle `@_callbacks` in a special way, using `transform_values` to create a new hash with new arrays for each child to independently add to.