Allow default symbol value to resolve to private method call
When a default value is a symbol and that symbol corresponds to a public method, OptStruct was already converting that symbol into the return value of the corresponding public method. However, it was not working for private methods. This commit uses `respond_to?` across all methods so that symbols for private/protected methods behave the same as public methods and the method is called to populate default values when present.